Create Dice_Roller
This creates 2 numbers added together like if it were dice. So you would run this code and the number that it prints out would be the sum of the 2 dice together so that is how many spaces you would move if this was a board game.


Going to work on creating a window that pops up with a button that you would click to get the number to make it easier to see instead of needing your terminal to see the number. 

Will also add animations later on once I figure out how to do it so that way I can animate the dice moving so you would not need to see the numbers that are summed up together.